+++ vyper/simple/algorithm/sort/bubble.vy
@@ -50,7 +50,7 @@ DirectionDescending: constant(uint8) = 2
 def main(_array: uint8[ARRAY_SIZE], direction: uint8) -> uint8[ARRAY_SIZE]:
     array: uint8[ARRAY_SIZE] = _array
     swapped: bool = True
-    for i: uint256 in range(ARRAY_SIZE - 1):
+    for i: uint8 in range(ARRAY_SIZE - 1):
         swapped = False
         for j: uint256 in range(ARRAY_SIZE - 1):
             if not j < ARRAY_SIZE - i - 1:
